Move `PauliBasis` and `ExtendedBasis` enums into `tqec.utils.enums`
