+++ src/tqec/__init__.py
@@ -3,3 +3,4 @@ from .compile import compile_block_graph as compile_block_graph
 from .computation.block_graph import BlockGraph as BlockGraph
 from .utils import Basis as Basis
 from .utils import NoiseModel as NoiseModel
+from .utils import PauliBasis as PauliBasis
