+++ src/tqec/plaquette/rpng/__init__.py
@@ -1,5 +1,3 @@
 from .rpng import RG as RG
 from .rpng import RPNG as RPNG
-from .rpng import ExtendedBasis as ExtendedBasis
-from .rpng import PauliBasis as PauliBasis
 from .rpng import RPNGDescription as RPNGDescription
