+++ src/tqec/utils/__init__.py
@@ -7,6 +7,8 @@ the other ``tqec`` sub-modules.
 """
 
 from .enums import Basis as Basis
+from .enums import ExtendedBasis as ExtendedBasis
 from .enums import Orientation as Orientation
+from .enums import PauliBasis as PauliBasis
 from .exceptions import TQECError as TQECError
 from .noise_model import NoiseModel as NoiseModel
